Implement navigation
This patch implements a nav-bar and the structure for the pages intended to be implemented.
This patch also implements a text-field element based on Tini Design System.
To reach the voting page, you now have to type a voting code (currently anything except about/guide) in the nav-bar and press join/Enter. You can also directly set the URL's pathname as the voting code, as sharing the voting URL is the intended way to share your voting with participants.
The voting page reads the voting code from the URL pathname. Future patches will implement fetching the voting data from server.