Adjust CandidateList sizing
candidate-container's width was overflowing the candidate-list's width, which was fixed by adding box-sizing: border-box;
candidate-list's height was set as 100%, which was unnecessary. This was removed.
These changes intend to move the CandidateList element torwards a more encapsulated form, allowing it to be manipulated by its parent page, instead of the current state, in which the page's layout is highly dependent on the CandidateList itself.