Implement button element
This patch implements a button element to replace the old buttons in the nav-bar.
Such button element has the properties 'reverse' and 'pill', which will allow us to use them in other places.
Tini Design System also specify the 'neutral' property for the button, which could be implemented in the future. This current implementation doesn't includes it, as we don't foresee its use in the near future.