Rename IServerProxy to IVotingProxy
Because this interface is intended to handle only the methods related to a currently running voting, not all server functionalities, its name should reflect this restriction.
Also, this interface is called a proxy because it represents an object from another adress space. I think it makes more sense that the name references the object itself (a voting) instead of the address space it's in (the server).