Instantiate the VotingProxy only once in each page
Previously, a new VotingProxy was instantiated on each render of the Voting/Results pages. This occured because the useRef hook was initialized with a call to the factory method that instantiates the proxy. This call was moved to the initializer function so that it will run only once for each page.